requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
tweepy
httpx
python-dotenv
